feature:bottom bar counts
So, its hard to track a site when the devs
inevitably change something.  As I was still
getting 200s but not getting any data pulls
because I hadn't built anything in to monitor
the counts as they were aggregating.  So now
the bottom count box is split into each of the
sites we visit.  Hosting a count for
how many homes found in a run.