feat:Updated results in terminal dash
After splitting up the dash to monitor each sites
returns.  Its hard to track when a site breaks
So, i added logic that if any data was returned
from a site, and the border hasn't already been
changed, to update it to a magenta color to
let me know that the site is still returning
data successfully.